Added Test Cases for DrinkAlgorithm, SalesService, and DrinkRepository. Made changes to each class in accordance to test results.

+++ Drink-Class-Library/Drink.cs
@@ -11,7 +11,7 @@
         public int price_level { get; private set; }
         public bool varies;
 
-        public int previousWeight;
+        public double previousWeight;
 
 
         public Drink(string drink_name, double price)
